feat: Add version export functionality
Introduces the ability to export conversation history to a file. This also includes enhancements to the header and version log UI to support this new feature, such as adding a new icon for export and integrating it into the header's toolbar. Minor refactoring of date handling in `ConversationPanel` to improve robustness.